refactor(RTL): promote RTL specs to replace their Enzyme counterparts
