+++ src/v2/Apps/Tag/__tests__/TagApp.jest.tsx
@@ -1,17 +1,18 @@
 import React from "react"
-import { Meta } from "react-head"
 import { TagAppFragmentContainer } from "../TagApp"
-import { setupTestWrapper } from "v2/DevTools/setupTestWrapper"
+import { setupTestWrapperTL } from "v2/DevTools/setupTestWrapper"
 import { graphql } from "react-relay"
 import { TagApp_Test_Query } from "v2/__generated__/TagApp_Test_Query.graphql"
 import { MockBoot } from "v2/DevTools"
+import { screen } from "@testing-library/react"
+import { findMetaTagBySelector } from "v2/DevTools"
 
 jest.unmock("react-relay")
 jest.mock("../Components/TagArtworkFilter", () => ({
   TagArtworkFilterRefetchContainer: () => <div />,
 }))
 
-const { getWrapper } = setupTestWrapper<TagApp_Test_Query>({
+const { renderWithRelay } = setupTestWrapperTL<TagApp_Test_Query>({
   Component: props => {
     return (
       <MockBoot>
@@ -31,42 +32,45 @@ const { getWrapper } = setupTestWrapper<TagApp_Test_Query>({
 
 describe("TagApp", () => {
   it("renders correctly", () => {
-    const wrapper = getWrapper({
+    renderWithRelay({
       Tag: () => ({
         name: "Example Tag",
       }),
     })
 
-    expect(wrapper.find("h1")).toHaveLength(1)
-    expect(wrapper.find("h1").text()).toEqual("Example Tag")
+    expect(screen.getByText("Example Tag")).toBeInTheDocument()
   })
 
-  it("renders meta description from query", () => {
-    const wrapper = getWrapper({
+  it("renders meta description from query", async () => {
+    renderWithRelay({
       Tag: () => ({
+        name: "Example Tag",
         description: "Tag Description",
       }),
     })
 
-    for (let i = 1; i <= 3; i++) {
-      expect(wrapper.find(Meta).at(i).prop("content")).toEqual(
-        "Tag Description"
-      )
-    }
+    const descriptionMeta = await findMetaTagBySelector(
+      "meta[name=description]"
+    )
+
+    expect(descriptionMeta).toHaveAttribute("content", "Tag Description")
   })
 
-  it("renders fallback meta description", () => {
-    const wrapper = getWrapper({
+  it("renders fallback meta description", async () => {
+    renderWithRelay({
       Tag: () => ({
         name: "Example",
         description: null,
       }),
     })
 
-    for (let i = 1; i <= 3; i++) {
-      expect(wrapper.find(Meta).at(i).prop("content")).toEqual(
-        "Browse all artworks with the Example tag on Artsy. Artsy has the largest collection of art on the Web; browse art by subject matter, medium, size and price."
-      )
-    }
+    const descriptionMeta = await findMetaTagBySelector(
+      "meta[name=description]"
+    )
+
+    expect(descriptionMeta).toHaveAttribute(
+      "content",
+      "Browse all artworks with the Example tag on Artsy. Artsy has the largest collection of art on the Web; browse art by subject matter, medium, size and price."
+    )
   })
 })
